Compromise proposal on the plugin base class
